feat: map sdk context libs로 이동 및 useMap 테스트 추가
