+++ frontend/src/libs/map-sdk/hooks/useMap.ts
@@ -1,0 +1,1 @@
+export { useMap } from '../contexts/KakaoMapContext';
